requirements.txt
Django==3.2.16
djangorestframework==3.12.4
djangorestframework-simplejwt==5.4.0
PyJWT==2.10.1
python-slugify-8.0.4
